Attempting to fix error
Initial error: Exception thrown: read access violation.
this->bullet was nullptr.

Check to see if bullet it built properly as this should not be empty/0
Added bullet.update(dt); into level
Player update has a pointer to bullet update, unsure if the code I added is needed because of this
Also added Bullet bullet; to level.h